% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_base.R
\name{ms.connect}
\alias{ms.connect}
\title{ms.connect}
\usage{
ms.connect(host, port = "3306", schema = NULL, user = connData$IAM_user,
  pass = connData$IAM_pass, ssl_ca_params = connData$db_mysql_ssl_ca_params,
  use_ssl = !is.null(connData$db_mysql_ssl_ca_params),
  jar = system.file("java", "mysql-connector-java-5.1.40-bin.jar", package =
  "mysqltools"))
}
